refactor: backwards compatibility for statements if right db version isnt available
