+++ apps/studio/components/interfaces/QueryPerformance/IndexAdvisor/IndexImprovementText.tsx
