+++ apps/studio/components/interfaces/QueryPerformance/QueryPerformance.utils.ts
@@ -24,3 +24,14 @@ export const formatDuration = (milliseconds: number) => {
 
   return parts.length > 0 ? parts.join(' ') : '0s'
 }
+
+export const transformLogsToJSON = (log: string) => {
+  try {
+    let jsonString = log.replace('[pg_stat_monitor] ', '')
+    jsonString = jsonString.replace(/""/g, '","')
+    const jsonObject = JSON.parse(jsonString)
+    return jsonObject
+  } catch (error) {
+    return null
+  }
+}
