+++ apps/studio/components/interfaces/QueryPerformance/QueryPerformanceChart.tsx
@@ -1,10 +1,10 @@
 import { useState, useMemo } from 'react'
 import { Tabs_Shadcn_, TabsContent_Shadcn_, TabsList_Shadcn_, TabsTrigger_Shadcn_ } from 'ui'
-import { QUERY_PERFORMANCE_CHART_TABS } from './QueryPerformanceChart.constants'
+import { QUERY_PERFORMANCE_CHART_TABS } from './QueryPerformance.constants'
 import { Loader2 } from 'lucide-react'
 import { ComposedChart } from 'components/ui/Charts/ComposedChart'
 import type { MultiAttribute } from 'components/ui/Charts/ComposedChart.utils'
-import type { ChartDataPoint } from '../QueryPerformanceData.utils'
+import type { ChartDataPoint } from './WithMonitor/WithMonitor.utils'
 
 interface QueryPerformanceChartProps {
   dateRange?: {
