--- apps/studio/components/interfaces/QueryPerformance/QueryPerformanceChart/QueryPerformanceChart.utils.ts
@@ -1,10 +1,0 @@
-export const transformLogsToJSON = (log: string) => {
-  try {
-    let jsonString = log.replace('[pg_stat_monitor] ', '')
-    jsonString = jsonString.replace(/""/g, '","')
-    const jsonObject = JSON.parse(jsonString)
-    return jsonObject
-  } catch (error) {
-    return null
-  }
-}
