+++ apps/studio/components/interfaces/QueryPerformance/hooks/useIsIndexAdvisorStatus.ts
@@ -1,4 +1,4 @@
-import { getIndexAdvisorExtensions } from 'components/interfaces/QueryPerformance/index-advisor.utils'
+import { getIndexAdvisorExtensions } from 'components/interfaces/QueryPerformance/IndexAdvisor/index-advisor.utils'
 import { useDatabaseExtensionsQuery } from 'data/database-extensions/database-extensions-query'
 import { useSelectedProjectQuery } from 'hooks/misc/useSelectedProject'
